[Add] improvements to wren math API
