+++ engine/bindings/private/wren_bindings.cpp
@@ -22,6 +22,7 @@
 #include "renderer/animation_bindings.hpp"
 #include "systems/lifetime_component.hpp"
 #include "time_module.hpp"
+#include "utility/math_bind.hpp"
 #include "utility/wren_entity.hpp"
 #include "wren_engine.hpp"
 
@@ -30,8 +31,7 @@
 
 namespace bindings
 {
-void BindMath(wren::ForeignModule& module);
-void BindMathHelper(wren::ForeignModule& module);
+
 void BindEntity(wren::ForeignModule& module);
 
 float TimeModuleGetDeltatime(TimeModule& self)
@@ -124,7 +124,6 @@ uint32_t GetEntity(WrenEntity& self) { return static_cast<uint32_t>(self.entity)
 void BindEngineAPI(wren::ForeignModule& module)
 {
     bindings::BindMath(module);
-    bindings::BindMathHelper(module);
     bindings::BindEntity(module);
 
     // Add modules here to expose them in scripting
@@ -214,157 +213,6 @@ void BindEngineAPI(wren::ForeignModule& module)
 
         transformClass.funcExt<bindings::TransformHelpersSetWorldTransform>("SetWorldTransform");
     }
-}
-
-// BINDING MATH TYPES
-
-namespace vector_ops
-{
-template <typename T>
-static T Default() { return {}; }
-
-template <typename T>
-static T Add(T& lhs, const T& rhs) { return lhs + rhs; }
-
-template <typename T>
-static T Sub(T& lhs, const T& rhs) { return lhs - rhs; }
-
-template <typename T>
-static T Neg(T& lhs) { return -lhs; }
-
-template <typename T>
-static T Mul(T& lhs, const T& rhs) { return lhs * rhs; }
-
-template <typename T>
-static bool Equals(T& lhs, const T& rhs) { return lhs == rhs; }
-
-template <typename T>
-static bool NotEquals(T& lhs, const T& rhs) { return lhs != rhs; }
-
-template <typename T>
-static T Normalized(T& v) { return glm::normalize(v); }
-
-template <typename T>
-static float Length(T& v) { return glm::length(v); }
-
-};
-
-class MathUtil
-{
-public:
-    static glm::vec3 ToEuler(glm::quat quat)
-    {
-        return glm::eulerAngles(quat);
-    }
-    static glm::quat ToQuat(glm::vec3 euler)
-    {
-        return glm::quat { euler };
-    }
-    static glm::vec3 ToDirectionVector(glm::quat quat)
-    {
-        return quat * glm::vec3(0.0f, 0.0f, -1.0f);
-    }
-    static glm::vec3 Mix(glm::vec3 start, glm::vec3 end, float t)
-    {
-        return glm::mix(start, end, t);
-    }
-    static float Dot(glm::vec3 a, glm::vec3 b)
-    {
-        return glm::dot(a, b);
-    }
-    static float Clamp(float a, float min, float max)
-    {
-        return glm::clamp(a, min, max);
-    }
-    static float Sqrt(float a)
-    {
-        return glm::sqrt(a);
-    }
-    static float Abs(float a)
-    {
-        return glm::abs(a);
-    }
-    static float PI()
-    {
-        return glm::pi<float>();
-    }
-    static float TwoPI()
-    {
-        return glm::two_pi<float>();
-    }
-    static float HalfPI()
-    {
-        return glm::half_pi<float>();
-    }
-
-    static float Distance(glm::vec3 pos1, glm::vec3 pos2)
-    {
-        return glm::distance(pos1, pos2);
-    }
-
-    static glm::vec3 Mul(glm::quat& lhs, const glm::vec3& rhs) { return lhs * rhs; }
-};
-
-template <typename T>
-void BindVectorTypeOperations(wren::ForeignKlassImpl<T>& klass)
-{
-    klass.template funcStaticExt<vector_ops::Default<T>>("Default");
-    klass.template funcExt<vector_ops::Add<T>>(wren::OPERATOR_ADD);
-    klass.template funcExt<vector_ops::Sub<T>>(wren::OPERATOR_SUB);
-    klass.template funcExt<vector_ops::Neg<T>>(wren::OPERATOR_NEG);
-    klass.template funcExt<vector_ops::Mul<T>>(wren::OPERATOR_MUL);
-    klass.template funcExt<vector_ops::Equals<T>>(wren::OPERATOR_EQUAL);
-    klass.template funcExt<vector_ops::NotEquals<T>>(wren::OPERATOR_NOT_EQUAL);
-    klass.template funcExt<vector_ops::Normalized<T>>("normalize");
-    klass.template funcExt<vector_ops::Length<T>>("length");
-}
-
-void bindings::BindMath(wren::ForeignModule& module)
-{
-    {
-        auto& vector2 = module.klass<glm::vec2>("Vec2");
-        vector2.ctor<float, float>();
-        vector2.var<&glm::vec2::x>("x");
-        vector2.var<&glm::vec2::y>("y");
-        BindVectorTypeOperations(vector2);
-    }
-
-    {
-        auto& vector3 = module.klass<glm::vec3>("Vec3");
-        vector3.ctor<float, float, float>();
-        vector3.var<&glm::vec3::x>("x");
-        vector3.var<&glm::vec3::y>("y");
-        vector3.var<&glm::vec3::z>("z");
-        BindVectorTypeOperations(vector3);
-    }
-
-    {
-        auto& quat = module.klass<glm::quat>("Quat");
-        quat.ctor<float, float, float, float>();
-        quat.var<&glm::quat::w>("w");
-        quat.var<&glm::quat::x>("x");
-        quat.var<&glm::quat::y>("y");
-        quat.var<&glm::quat::z>("z");
-        BindVectorTypeOperations(quat);
-        quat.funcExt<MathUtil::Mul>("mul");
-    }
-}
-
-void bindings::BindMathHelper(wren::ForeignModule& module)
-{
-    auto& mathUtilClass = module.klass<MathUtil>("Math");
-    mathUtilClass.funcStatic<&MathUtil::ToEuler>("ToEuler");
-    mathUtilClass.funcStatic<&MathUtil::ToDirectionVector>("ToVector");
-    mathUtilClass.funcStatic<&MathUtil::ToQuat>("ToQuat");
-    mathUtilClass.funcStatic<&MathUtil::Mix>("Mix");
-    mathUtilClass.funcStatic<&MathUtil::Dot>("Dot");
-    mathUtilClass.funcStatic<&MathUtil::Clamp>("Clamp");
-    mathUtilClass.funcStatic<&MathUtil::Sqrt>("Sqrt");
-    mathUtilClass.funcStatic<&MathUtil::Abs>("Abs");
-    mathUtilClass.funcStatic<&MathUtil::PI>("PI");
-    mathUtilClass.funcStatic<&MathUtil::TwoPI>("TwoPI");
-    mathUtilClass.funcStatic<&MathUtil::HalfPI>("HalfPI");
-    mathUtilClass.funcStatic<&MathUtil::Distance>("Distance");
 }
 
 void bindings::BindEntity(wren::ForeignModule& module)
